fix: np img

+++ test/test_artefacts.py
@@ -6,7 +6,7 @@ from doctr.documents import DocumentFile
 
 def test_qr_code_detector(mock_image_folder):
     detector = QRCodeDetector()
-    image = np.random.randint(low=0, high=255, size=(512, 512, 3))
+    image = np.random.randint(low=0, high=255, size=(512, 512, 3), dtype=np.uint8)
     qrcode = detector(image)
     assert qrcode is None
